[SYCL] Move `codeToString` `common.hpp` -> `exception.hpp`
That is its only use inside includes with others under `source/`. Breaks
cyclical include dependency between those headers.
