+++ sycl/source/exception.cpp
@@ -94,5 +94,24 @@ std::error_code make_error_code(sycl::errc Err) noexcept {
   return {static_cast<int>(Err), sycl_category()};
 }
 
+namespace detail {
+const char *stringifyErrorCode(pi_int32 error) {
+  switch (error) {
+#define _PI_ERRC(NAME, VAL)                                                    \
+  case NAME:                                                                   \
+    return #NAME;
+#define _PI_ERRC_WITH_MSG(NAME, VAL, MSG)                                      \
+  case NAME:                                                                   \
+    return MSG;
+#include <sycl/detail/pi_error.def>
+#undef _PI_ERRC
+#undef _PI_ERRC_WITH_MSG
+
+  default:
+    return "Unknown error code";
+  }
+}
+} // namespace detail
+
 } // namespace _V1
 } // namespace sycl
