+++ sycl/source/spirv_ops.cpp
@@ -8,6 +8,7 @@
 
 #include <CL/__spirv/spirv_ops.hpp>
 #include <detail/platform_util.hpp>
+#include <sycl/detail/iostream_proxy.hpp>
 #include <sycl/exception.hpp>
 
 #include <atomic>
